Moved Config.java to the config folder
Created a config folder after the Config.java was already being used by multiple people so didn't move it until they were finished.
