+++ src/main/java/frc/robot/Robot.java
@@ -8,6 +8,7 @@ import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.commands.SetModuleEncoderFromLamprey;
+import frc.robot.config.Config;
 
 /**
  * The VM is configured to automatically run this class, and to call the
