+++ src/main/java/frc/robot/RobotContainer.java
@@ -9,9 +9,8 @@ import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.controller.PIDController;
 import edu.wpi.first.wpilibj.controller.ProfiledPIDController;
-import frc.robot.Config.AutoConstants;
-import frc.robot.Config.DriveConstants;
-import frc.robot.Config.OIConstants;
+import frc.robot.config.Config.*;
+import frc.robot.config.Config;
 import frc.robot.subsystems.DriveSubsystem;
 import edu.wpi.first.wpilibj.geometry.Pose2d;
 import edu.wpi.first.wpilibj.geometry.Rotation2d;
