+++ src/main/java/frc/robot/subsystems/SwerveModule.java
@@ -12,7 +12,7 @@ import com.revrobotics.ControlType;
 
 import edu.wpi.first.wpilibj.AnalogPotentiometer;
 import edu.wpi.first.wpilibj.DriverStation;
-import frc.robot.Config;
+import frc.robot.config.Config;
 
 import edu.wpi.first.wpilibj.geometry.Rotation2d;
 import edu.wpi.first.wpilibj.kinematics.SwerveModuleState;
